src/stock_cli/data/system_monitor.py
"""
系统监控模块
"""

import datetime as dt
from typing import Any, Dict

try:
    import psutil

    PSUTIL_OK = True
except ImportError:
    PSUTIL_OK = False


class SystemMonitor:
    """系统资源监控"""

    @staticmethod
    def get_system_info() -> Dict[str, Any]:
        """获取系统基础信息"""
        try:
            if PSUTIL_OK:
                cpu_percent = psutil.cpu_percent(interval=0.1)
                memory = psutil.virtual_memory()
                disk = psutil.disk_usage("/")
                net_io = psutil.net_io_counters()
                boot_time = psutil.boot_time()

                return {
                    "CPU使用率": f"{cpu_percent:.1f}%",
                    "内存使用": f"{memory.percent:.1f}%",
                    "内存总量": f"{memory.total / 1024 / 1024 / 1024:.1f}GB",
                    "内存可用": f"{memory.available / 1024 / 1024 / 1024:.1f}GB",
                    "磁盘使用": f"{disk.percent:.1f}%",
                    "磁盘总量": f"{disk.total / 1024 / 1024 / 1024:.1f}GB",
                    "磁盘可用": f"{disk.free / 1024 / 1024 / 1024:.1f}GB",
                    "网络发送": f"{net_io.bytes_sent / 1024 / 1024:.1f}MB",
                    "网络接收": f"{net_io.bytes_recv / 1024 / 1024:.1f}MB",
                    "启动时间": dt.datetime.fromtimestamp(boot_time).strftime(
                        "%m-%d %H:%M"
                    ),
                    "当前时间": dt.datetime.now().strftime("%H:%M:%S"),
                }
            else:
                # 模拟数据当psutil不可用时
                import random

                memory_percent = random.uniform(40, 80)
                disk_percent = random.uniform(30, 70)
                return {
                    "CPU使用率": f"{random.uniform(10, 50):.1f}%",
                    "内存使用": f"{memory_percent:.1f}%",
                    "内存总量": "16.0GB",
                    "内存可用": f"{16 * (100 - memory_percent) / 100:.1f}GB",
                    "磁盘使用": f"{disk_percent:.1f}%",
                    "磁盘总量": "1024.0GB",
                    "磁盘可用": f"{1024 * (100 - disk_percent) / 100:.1f}GB",
                    "网络发送": "0.0MB",
                    "网络接收": "0.0MB",
                    "启动时间": "08-08 09:00",
                    "当前时间": dt.datetime.now().strftime("%H:%M:%S"),
                }
        except Exception:
            return {
                "CPU使用率": "N/A",
                "内存使用": "N/A",
                "磁盘使用": "N/A",
                "内存总量": "N/A",
                "内存可用": "N/A",
                "磁盘总量": "N/A",
                "磁盘可用": "N/A",
                "网络发送": "N/A",
                "网络接收": "N/A",
                "启动时间": "N/A",
                "当前时间": dt.datetime.now().strftime("%H:%M:%S"),
            }
