pyproject.toml
[project]
name = "stock-cli"
version = "1.0.0"
description = "A Task-Driven ReAct Stock Analysis Agent with CLI Interface"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "chromadb>=0.4.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "PyYAML>=6.0",
    "pyfiglet>=0.8.post1",
    "mcp>=1.12.4",
    "akshare>=1.17.31",
    "openai>=1.99.6",
    "prompt-toolkit>=3.0.51",
]

[project.scripts]
stock-cli = "stock_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_cli"]

[tool.uv]
package = true
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"



dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.4.0",
    "ruff>=0.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]
